Refactored old gauge scripts into firstweek setup and ff.
Removed checkGaugeRewards and exampleSetup for new first
and sendond week scripts. First week sets gauge weights
and rewards. Second week fast forwards one week and checks
claimable rewards.
